add partialBuyFixedRateBond functionality!

+++ src/main/java/fund/data/assets/utils/FinancialAndAnotherConstants.java
@@ -6,7 +6,7 @@ package fund.data.assets.utils;
  * @author MarkDementev a.k.a JavaMarkDem
  */
 public class FinancialAndAnotherConstants {
-    public static final int YEAR_DAYS_COUNT = 365;
-    public static final int LEAP_YEAR_DAYS_COUNT = 366;
+    public static final float YEAR_DAYS_COUNT = 365.0F;
+    public static final float LEAP_YEAR_DAYS_COUNT = 366.0F;
     public static final float RUSSIAN_TAX_SYSTEM_CORRECTION_VALUE = 0.87F;
 }
